requirements.txt
# AURA Healthcare Framework - Backend Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
aiofiles==23.2.1

# Database
pymongo==4.6.0
motor==3.3.2
sqlalchemy==2.0.23

# Data Models
pydantic==2.5.0
pydantic-settings==2.1.0

# AI & ML Core
langchain==0.0.350
langchain-community==0.0.10
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
openai==1.3.7

# Vector Database
qdrant-client==1.6.9

# NLP & Text Processing
spacy==3.7.2
nltk==3.8.1

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
docx2txt==0.8

# Data Processing
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Utilities
httpx==0.25.2
python-dotenv==1.0.0
requests==2.31.0
redis==5.0.1

# Development Tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
